+++ projects/angular2-draggable/src/lib/angular-resizable.directive.ts
@@ -8,7 +8,7 @@ import {
 import { HelperBlock } from './widgets/helper-block';
 import { ResizeHandle } from './widgets/resize-handle';
 import { ResizeHandleType } from './models/resize-handle-type';
-import { Position } from './models/position';
+import { Position, IPosition } from './models/position';
 import { Size } from './models/size';
 import { IResizeEvent } from './models/resize-event';
 
@@ -36,6 +36,9 @@ export class AngularResizableDirective implements OnInit, OnChanges, OnDestroy, 
   /** Initial Size and Position */
   private _initSize: Size = null;
   private _initPos: Position = null;
+
+  /** Snap to gird */
+  private _gridSize: IPosition = null;
 
   private _bounding: any = null;
 
@@ -79,6 +82,24 @@ export class AngularResizableDirective implements OnInit, OnChanges, OnDestroy, 
    *  String: Possible values: "parent".
    */
   @Input() rzContainment: string | HTMLElement = null;
+
+  /**
+   * Snaps the resizing element to a grid, every x and y pixels.
+   * A number for both width and height or an array values like [ x, y ]
+   */
+  @Input() rzGrid: number | number[] = null;
+
+  /** The minimum width the resizable should be allowed to resize to. */
+  @Input() rzMinWidth: number = null;
+
+  /** The minimum height the resizable should be allowed to resize to. */
+  @Input() rzMinHeight: number = null;
+
+  /** The maximum width the resizable should be allowed to resize to. */
+  @Input() rzMaxWidth: number = null;
+
+  /** The maximum height the resizable should be allowed to resize to. */
+  @Input() rzMaxHeight: number = null;
 
   /** emitted when start resizing */
   @Output() rzStart = new EventEmitter<IResizeEvent>();
@@ -276,6 +297,7 @@ export class AngularResizableDirective implements OnInit, OnChanges, OnDestroy, 
       if (this._containment) {
         this.getBounding();
       }
+      this.getGridSize();
       this.startResize(handle);
     }
   }
@@ -341,10 +363,13 @@ export class AngularResizableDirective implements OnInit, OnChanges, OnDestroy, 
   private resizeTo(p: Position) {
     p.subtract(this._origMousePos);
 
+    const tmpX = Math.round(p.x / this._gridSize.x) * this._gridSize.x;
+    const tmpY = Math.round(p.y / this._gridSize.y) * this._gridSize.y;
+
     if (this._handleResizing.type.match(/n/)) {
       // n, ne, nw
-      this._currPos.y = this._origPos.y + p.y;
-      this._currSize.height = this._origSize.height - p.y;
+      this._currPos.y = this._origPos.y + tmpY;
+      this._currSize.height = this._origSize.height - tmpY;
 
       // check bounds
       if (this._containment) {
@@ -363,7 +388,7 @@ export class AngularResizableDirective implements OnInit, OnChanges, OnDestroy, 
       this.adjustByRatio('h');
     } else if (this._handleResizing.type.match(/s/)) {
       // s, se, sw
-      this._currSize.height = this._origSize.height + p.y;
+      this._currSize.height = this._origSize.height + tmpY;
 
       // aspect ratio
       this.adjustByRatio('h');
@@ -371,14 +396,14 @@ export class AngularResizableDirective implements OnInit, OnChanges, OnDestroy, 
 
     if (this._handleResizing.type.match(/e/)) {
       // e, ne, se
-      this._currSize.width = this._origSize.width + p.x;
+      this._currSize.width = this._origSize.width + tmpX;
 
       // aspect ratio
       this.adjustByRatio('w');
     } else if (this._handleResizing.type.match(/w/)) {
       // w, nw, sw
-      this._currSize.width = this._origSize.width - p.x;
-      this._currPos.x = this._origPos.x + p.x;
+      this._currSize.width = this._origSize.width - tmpX;
+      this._currPos.x = this._origPos.x + tmpX;
 
       // check bounds
       if (this._containment) {
@@ -398,6 +423,7 @@ export class AngularResizableDirective implements OnInit, OnChanges, OnDestroy, 
     }
 
     this.checkBounds();
+    this.checkSize();
     this.doResize();
   }
 
@@ -421,7 +447,6 @@ export class AngularResizableDirective implements OnInit, OnChanges, OnDestroy, 
 
   private checkBounds() {
     if (this._containment) {
-      const container = this._containment;
       const maxWidth = this._bounding.width - this._bounding.pr - this.el.nativeElement.offsetLeft;
       const maxHeight = this._bounding.height - this._bounding.pb - this.el.nativeElement.offsetTop;
 
@@ -432,6 +457,24 @@ export class AngularResizableDirective implements OnInit, OnChanges, OnDestroy, 
       if (this._currSize.height > maxHeight) {
         this._currSize.height = maxHeight;
       }
+    }
+  }
+
+  private checkSize() {
+    if (this.rzMaxWidth && this._currSize.width > this.rzMaxWidth) {
+      this._currSize.width = this.rzMaxWidth;
+    }
+
+    if (this.rzMinWidth && this._currSize.width < this.rzMinWidth) {
+      this._currSize.width = this.rzMinWidth;
+    }
+
+    if (this.rzMaxHeight && this._currSize.height > this.rzMaxHeight) {
+      this._currSize.height = this.rzMaxHeight;
+    }
+
+    if (this.rzMinHeight && this._currSize.height < this.rzMinHeight) {
+      this._currSize.height = this.rzMinHeight;
     }
   }
 
@@ -459,5 +502,18 @@ export class AngularResizableDirective implements OnInit, OnChanges, OnDestroy, 
       this.renderer.setStyle(this._containment, 'position', 'relative');
     }
     this._bounding = null;
+  }
+
+  private getGridSize() {
+    // set default value:
+    this._gridSize = { x: 1, y: 1 };
+
+    if (this.rzGrid) {
+      if (typeof this.rzGrid === 'number') {
+        this._gridSize = { x: this.rzGrid, y: this.rzGrid };
+      } else if (Array.isArray(this.rzGrid)) {
+        this._gridSize = { x: this.rzGrid[0], y: this.rzGrid[1] };
+      }
+    }
   }
 }